Move Json-utils to Json-package
