+++ src/main/kotlin/no/nav/eessi/pensjon/journalforing/services/oppgave/OppgaveService.kt
@@ -2,7 +2,7 @@ package no.nav.eessi.pensjon.journalforing.services.oppgave
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import no.nav.eessi.pensjon.journalforing.metrics.counter
-import no.nav.eessi.pensjon.journalforing.utils.mapAnyToJson
+import no.nav.eessi.pensjon.journalforing.json.mapAnyToJson
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpMethod
